added controller

+++ KK/src/main/java/com/knowledgeknockout/kk/interfaces/QuizCardJDBC.java
@@ -3,5 +3,5 @@ package com.knowledgeknockout.kk.interfaces;
 import com.knowledgeknockout.kk.entity.QuizCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizCardJDBC extends JpaRepository<QuizCard, Integer> {
+public interface QuizCardJDBC extends JpaRepository<QuizCard, Long> {
 }
